+++ src/Client/public/results.js
@@ -51,8 +51,11 @@ function getL10n() {
         team: l ("Team", "Команда", "Команда"),
         points: l ("Points", "Очки", "Очки"),
         place: l ("Place", "Место", "Місце"),
+        rating: l ("Rating", "Рейтинг", "Рейтинг"),
+        questionRating: l ("Question's rating", "Рейтинг вопроса", "Рейтинг питання"),
         toShortView: l ('less details', "скрыть детали", "сховати подробицi"),
         toFullView: l ('more details', "показать детали", "показати подробицi"),
+        tour: l ('Tour', "Тур", "Тур")
     }
 
     return l10n
@@ -63,10 +66,24 @@ function shortTable(data, l10n)
     var $topContainer = $('<div/>', {'style': 'display: grid'})
     var $container = $('<div/>', {'class': 'table-container'}).appendTo($topContainer)
 
+    var toursCount = data.Questions.filter(function (obj) {return obj.EOT}).length
+
+    if (data.Questions.length > 0 && !data.Questions[data.Questions.length - 1].EOT){
+        toursCount = toursCount + 1
+    }
+
     var $table = $('<table>', {'class': 'table is-hoverable'}).appendTo($container)
-    $table
-    .append('<thead>').children('thead')
-    .append('<tr/>').children('tr').append('<th>#</th><th>'+l10n.team+'</th><th>'+l10n.points+'</th><th>'+l10n.place+'</th>')
+    var $hRow = $table.append('<thead>').children('thead').append('<tr/>').children('tr')
+
+    $('<th>', {'text': '#'}).appendTo($hRow)
+    $('<th>', {'text': l10n.team}).appendTo($hRow)
+    for (var i = 0; i < toursCount; i++){
+        $('<th>', {'text': l10n.tour + " " + (i + 1), 'style': 'white-space: nowrap; text-align: center'}).appendTo($hRow)
+    }
+    $('<th>', {'text': l10n.points, 'style': 'text-align: center'}).appendTo($hRow)
+    $('<th>', {'text': l10n.rating, 'style': 'text-align: center'}).appendTo($hRow)
+    $('<th>', {'text': l10n.place, 'style': 'text-align: center'}).appendTo($hRow)
+
 
     var $body = $table.append('<tbody>').children('tbody')
 
@@ -77,7 +94,13 @@ function shortTable(data, l10n)
             .append('<td>'+team.TeamId+'</td>')
             .append('<td>'+team.TeamName+'</td>');
 
+        for (var j = 0; j < toursCount; j++){
+            var tour = team.Tours[j]
+            $('<td>', {'text': tour.Points, 'style': 'text-align: center'}).appendTo($row)
+        }
+
         $('<td>', {'text': team.Points, 'style': 'text-align: center'}).appendTo($row)
+        $('<td>', {'text': team.Rating, 'style': 'text-align: center'}).appendTo($row)
         $('<td>', {'text': (team.PlaceFrom == team.PlaceTo) ? team.PlaceFrom : team.PlaceFrom + '-' + team.PlaceTo, 'style': 'text-align: center'}).appendTo($row)
         $body.append($row)
     }
@@ -99,9 +122,25 @@ function fullTable(data, l10n)
         $hRow.append('<th>'+qw.Name+'</th>')
     }
     $hRow.append('<th>'+l10n.points+'</th>')
+    $hRow.append('<th>'+l10n.rating+'</th>')
     $hRow.append('<th>'+l10n.place+'</th>')
 
     var $tbody = $('<tbody>').appendTo($table)
+
+    // question's rating
+    var $row = $('<tr/>');
+    $('<td>',  {'style': 'border-bottom: double;'}).appendTo($row)
+    $('<th>', {'text': l10n.questionRating, 'style': 'white-space: nowrap; font-weight: normal; font-style: italic; border-bottom: double;'}).appendTo($row)
+    for (var j = 0; j<data.Questions.length; j++){
+        $('<td>', {'text': data.Questions[j].Rating, 'style': 'font-style: italic; border-bottom: double;'}).appendTo($row)
+    }
+    $('<td>',  {'style': 'border-bottom: double;'}).appendTo($row)
+    $('<td>',  {'style': 'border-bottom: double;'}).appendTo($row)
+    $('<td>',  {'style': 'border-bottom: double;'}).appendTo($row)
+
+    $tbody.append($row)
+
+
     for (var i = 0; i < data.Teams.length; i++){
         var team = data.Teams[i]
 
@@ -114,10 +153,13 @@ function fullTable(data, l10n)
             var key = JSON.stringify(qw.Key)
             var result = team.Details[key]
             var txt = result ? result : ""
-            $row.append('<td>'+txt+'</td>')
+            var style = ""
+            if (qw.EOT) {style = 'border-right: dotted;'}
+            $('<td>',  {'text': txt, 'style': style}).appendTo($row)
         }
 
         $('<td>', {'text': team.Points, 'style': 'text-align: center'}).appendTo($row)
+        $('<td>', {'text': team.Rating, 'style': 'text-align: center'}).appendTo($row)
         $('<td>', {'text': (team.PlaceFrom == team.PlaceTo) ? team.PlaceFrom : team.PlaceFrom + '-' + team.PlaceTo, 'style': 'text-align: center'}).appendTo($row)
 
         $tbody.append($row)
@@ -160,7 +202,7 @@ $(document).ready(function() {
     var settings = $.extend({}, defaults, parameters);
 
     var l10n = getL10n()
-    //alert(JSON.stringify(l10n))
+    // alert(JSON.stringify(l10n))
 
     var title = settings.quizName + " - " + l10n.results
     document.title = title
@@ -179,7 +221,7 @@ $(document).ready(function() {
         $(".hero-foot").show()
     }
 
-    var url = "https://www.open-quiz.com/static/" + settings.quiz + "-" + settings.token + "/results.json"
+    var url = "https://www.open-quiz.com/static/" + settings.quiz + "-" + settings.token + "/results.json?nocache=" + (new Date()).getTime()
     $.getJSON(url, function(data) {displayResults(data, settings.teamId, l10n)});
 });